[board-server] Remove invite mechanism
Fixes #4896
