+++ packages/board-server/src/server/boards/index.ts
@@ -15,8 +15,6 @@ import getBoard from "./get.js";
 import post from "./post.js";
 import invokeBoard from "./invoke.js";
 import describeBoard from "./describe.js";
-import inviteList from "./invite-list.js";
-import inviteUpdate from "./invite-update.js";
 import runBoard from "./run.js";
 import handleAssetsDriveRequest from "./assets-drive.js";
 import { loadBoard, parseBoardId } from "./loader.js";
@@ -49,18 +47,6 @@ export function serveBoardsAPI(serverConfig: ServerConfig): Router {
     parseBoardId({ addJsonSuffix: true }),
     describeBoard
   );
-  router.get(
-    "/@:user/:name.invite",
-    requireAuth(),
-    parseBoardId({ addJsonSuffix: true }),
-    inviteList
-  );
-  router.post(
-    "/@:user/:name.invite",
-    requireAuth(),
-    parseBoardId({ addJsonSuffix: true }),
-    inviteUpdate
-  );
 
   // Non suffixed routes. These routes treat the board name as an opaque
   // identifier and do not try to deconstruct it. The board is loaded from
@@ -75,18 +61,6 @@ export function serveBoardsAPI(serverConfig: ServerConfig): Router {
     runBoard(serverConfig, req, res)
   );
   router.post("/@:user/:name/describe", parseBoardId(), describeBoard);
-  router.get(
-    "/@:user/:name/invites",
-    requireAuth(),
-    parseBoardId(),
-    inviteList
-  );
-  router.post(
-    "/@:user/:name/invites",
-    requireAuth(),
-    parseBoardId(),
-    inviteUpdate
-  );
 
   router.post(
     "/@:user/:name/assets/drive/:driveId",
