+++ packages/board-server/src/server/store.ts
@@ -10,14 +10,6 @@ import { FirestoreStorageProvider } from "./storage-providers/firestore.js";
 export const EXPIRATION_TIME_MS = 1000 * 60 * 60 * 24 * 2; // 2 days
 export const INVITE_EXPIRATION_TIME_MS = 1000 * 60 * 60 * 24 * 4; // 4 days
 
-export type GetUserStoreResult =
-  | { success: true; store: string }
-  | { success: false; error: string };
-
-export type OperationResult =
-  | { success: true }
-  | { success: false; error: string };
-
 export function getStore(): FirestoreStorageProvider {
   const db = process.env["FIRESTORE_DB_NAME"] || "board-server";
   return new FirestoreStorageProvider(db);
